pyproject.toml
[tool.poetry]
name = "notes-django"
version = "0.1.0"
description = ""
authors = ["Max <gladkovnissan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2"
psycopg2 = "^2.9.10"
python-dotenv = "^1.1.0"
pillow = "^11.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
